app/config/gimpbaseconfig.h use gint64 for all memsize properties.
2003-11-16  Sven Neumann  <sven@gimp.org>

	* app/config/gimpbaseconfig.h
	* app/config/gimpcoreconfig.h: use gint64 for all memsize properties.

	* app/base/tile-manager.[ch] (tile_manager_get_memsize): since
	tiles can be swapped out, a tilemanager can be larger than gsize
	(on 32bit platforms). Use a gint64 to avoid an overflow.

	* app/core/gimp.c
	* app/core/gimpbrush.c
	* app/core/gimpbrushpipe.c
	* app/core/gimpbuffer.c
	* app/core/gimpchannel.c
	* app/core/gimpcontainer.c
	* app/core/gimpcontext.c
	* app/core/gimpdata.c
	* app/core/gimpdatafactory.c
	* app/core/gimpdrawable.c
	* app/core/gimpgradient.c
	* app/core/gimpimage-undo-push.c
	* app/core/gimpimage-undo.[ch]
	* app/core/gimpimage.c
	* app/core/gimpitem.c
	* app/core/gimpitemundo.[ch]
	* app/core/gimplayer.c
	* app/core/gimplist.c
	* app/core/gimpobject.[ch]
	* app/core/gimppalette.c
	* app/core/gimpparasitelist.c
	* app/core/gimppattern.c
	* app/core/gimpundo.[ch]
	* app/core/gimpundostack.c
	* app/core/gimpviewable.c
	* app/text/gimptext.c
	* app/text/gimptextlayer.c
	* app/vectors/gimpstroke.c
	* app/vectors/gimpvectors.c: use gint64 for gimp_object_get_memsize()
	and all its implementations.

	* app/display/gimpdisplayshell-title.c
	* app/gui/info-window.c: changed accordingly.

+++ app/base/tile-manager.h
@@ -19,7 +19,7 @@
 #ifndef __TILE_MANAGER_H__
 #define __TILE_MANAGER_H__
 
-struct _PixelDataHandle 
+struct _PixelDataHandle
 {
   guchar                 *data;
   gint                    width;
@@ -56,7 +56,7 @@ void          tile_manager_unref             (TileManager *tm);
 void          tile_manager_set_validate_proc (TileManager      *tm,
 					      TileValidateProc  proc);
 
-/* Get a specified tile from a tile manager. 
+/* Get a specified tile from a tile manager.
  */
 Tile        * tile_manager_get_tile          (TileManager *tm,
 					      gint         xpixel,
@@ -98,7 +98,7 @@ void          tile_invalidate                (Tile        **tile_ptr,
 					      TileManager  *tm,
 					      gint          tile_num);
 void          tile_invalidate_tile           (Tile        **tile_ptr,
-					      TileManager  *tm, 
+					      TileManager  *tm,
 					      gint          xpixel,
 					      gint          ypixel);
 
@@ -123,7 +123,7 @@ void          tile_manager_set_offsets       (TileManager       *tm,
 					      gint               x,
 					      gint               y);
 
-gsize         tile_manager_get_memsize       (const TileManager *tm);
+gint64        tile_manager_get_memsize       (const TileManager *tm);
 
 void          tile_manager_get_tile_coordinates (TileManager *tm,
 						 Tile        *tile,
